+++ vscode-extension/src/lib/types/RangeToShow.ts
@@ -1,0 +1,5 @@
+export enum RangeToShow {
+    ManifestRanges = "Manifest Ranges",
+    LogicalRanges = "Logical Ranges",
+    ObjectRanges = "Object Ranges"
+}